Changes committed in PrimitivesTest

+++ src/main/java/PrimitivesTest.java
@@ -1,7 +1,7 @@
 public class PrimitivesTest {
     public static <Int> void main(String[] args) {
-        int b = 34;
-        int s = -34;
+        byte b = 34;
+        short s = -34;
         int i = 34;
         long l = 65L;
         float myValue = 12;
